Kun rendre SyntaxHighlighteren på client side
Ref PR #226, så prøver vi dette for å fikse https://github.com/react-syntax-highlighter/react-syntax-highlighter/issues/538.
