+++ web/app/features/teaser/TeaserPage.tsx
@@ -1,18 +1,10 @@
-import { useEffect, useState } from 'react'
 import Countdown, { CountdownRendererFn } from 'react-countdown'
 import { Link } from '@remix-run/react'
 
 import { PostPreview } from '../post-preview/PostPreview'
 
 import { LinkToArchive } from '~/components/LinkToArchive'
-
-const useClientSideOnly = () => {
-  const [isClientSide, setIsClientSide] = useState(false)
-  useEffect(() => {
-    setIsClientSide(true)
-  }, [])
-  return isClientSide
-}
+import useClientSideOnly from '~/hooks/useClientSideOnly'
 
 interface CountdownRendererFnProps {
   days: number
