+++ web/app/hooks/useClientSideOnly.ts
@@ -1,0 +1,11 @@
+import { useEffect, useState } from 'react'
+
+const useClientSideOnly = () => {
+  const [isClientSide, setIsClientSide] = useState(false)
+  useEffect(() => {
+    setIsClientSide(true)
+  }, [])
+  return isClientSide
+}
+
+export default useClientSideOnly
